python/pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "legalize"
version = "0.1.0"
description = "Official Python client for the Legalize API"
readme = "README.md"
license = { file = "../LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "Legalize", email = "dev@legalize.dev" }]
keywords = ["legalize", "legal", "laws", "api", "sdk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "pytest-httpx>=0.30",
    "hypothesis>=6.100",
    "vcrpy>=6",
    "schemathesis>=3.30",
    "ruff>=0.6",
    "mypy>=1.11",
    "datamodel-code-generator>=0.26",
    "bandit>=1.7",
]

[project.urls]
Homepage = "https://legalize.dev"
Documentation = "https://legalize.dev/api/docs"
Repository = "https://github.com/legalize-dev/legalize-sdks"
Issues = "https://github.com/legalize-dev/legalize-sdks/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/legalize"]

[tool.hatch.build.targets.sdist]
include = ["src/legalize", "README.md", "../LICENSE"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",    # pycodestyle + pyflakes
    "I",              # isort
    "B",              # bugbear
    "UP",             # pyupgrade
    "N",              # pep8-naming
    "S",              # bandit
    "RUF",            # ruff rules
]
ignore = [
    "S101",  # assert in tests
    "E501",  # line length handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S105", "S106", "S311"]
"examples/**" = ["S105", "S106", "T201"]
"src/legalize/models/**" = ["N815", "UP"]  # generated code

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
warn_unused_ignores = true
show_error_codes = true
exclude = ["tests/", "examples/", "src/legalize/models/"]

[[tool.mypy.overrides]]
module = "legalize.models.*"
ignore_errors = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "unit: fast unit tests with mocked HTTP",
    "webhooks: webhook signature vectors",
    "property: Hypothesis property-based tests",
    "contract: schemathesis contract tests",
    "vcr: VCR-replayed tests against recorded cassettes",
    "integration: tests against a live staging API",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:vcr",
]

[tool.coverage.run]
source = ["src/legalize"]
branch = true
omit = [
    "src/legalize/models/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 95
show_missing = true
